‘literate-elisp-src-ids-loop’ is now aware of source buffers
Aha! This should fix the failed ERT test (glad I put that in!).
Now the ‘literate-elisp-src-ids-loop’ macro takes a second argument
before the body, BUF, which specifies the relevant source buffer.
(Usually passing it ‘in’ or ‘buf’ will do the trick, depending on the
arguments of the function calling the macro.)